[feat ] Remove dependency on youtube-search package
