+++ src/lib/constants.ts
@@ -1,3 +1,4 @@
 export const API_HOST = process.env.API_HOST
 export const YOUTUBE_KEY = process.env.YOUTUBE_KEY
 export const WS_HOST = process.env.WS_HOST
+export const YOUTUBE_API = 'https://www.googleapis.com/youtube/v3/search'
